# api-docs.json
{"openapi":"3.1.0","info":{"title":"Simulador de Investimentos API","description":"API para consultar ativos e simular investimentos","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"UserController","description":"Operações relacionadas à manipulação de usuários"},{"name":"AssetController","description":"Operações relacionadas à consulta de ativos"},{"name":"OrderController","description":"Operações relacionadas à compra, venda e consulta de ordens de ativos"},{"name":"PositionController","description":"Operações relacionadas às posições de ativos nas carteiras"},{"name":"WalletController","description":"Operações relacionadas às carteiras dos usuários"}],"paths":{"/api/v1/orders/sell/{userId}":{"put":{"tags":["OrderController"],"summary":"Efetua venda de um ativo","description":"Retorna as informações completas da venda efetuada","operationId":"sellAsset","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellOrderDTO"}}},"required":true},"responses":{"200":{"description":"Venda efetuada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}},"400":{"description":"Quantidade de ações insuficiente para venda"},"404":{"description":"Usuário, ordem ou posição não encontrada"},"500":{"description":"Erro interno"}}}},"/api/v1/users":{"get":{"tags":["UserController"],"summary":"Retorna todos os usuários","description":"Retorna uma lista com todos os usuários cadastrados","operationId":"getAllUsers","responses":{"200":{"description":"Usuários retornados com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDTO"}}}}},"500":{"description":"Erro interno"}}},"post":{"tags":["UserController"],"summary":"Criar um novo usuário","description":"Cria um novo usuário com saldo de R$1.000,00","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}},"required":true},"responses":{"201":{"description":"Usuário criado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}},"500":{"description":"Erro interno"}}}},"/api/v1/orders/buy/{userId}":{"post":{"tags":["OrderController"],"summary":"Efetua compra de um ativo","description":"Retorna as informações completas da compra efetuada","operationId":"buyAsset","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyRequestDTO"}}},"required":true},"responses":{"200":{"description":"Compra efetuada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}},"400":{"description":"Saldo insuficiente ou quantidade de ações inválida"},"404":{"description":"Usuário ou ativo não encontrado"},"500":{"description":"Erro interno"}}}},"/api/v1/wallets/{id}":{"get":{"tags":["WalletController"],"summary":"Consultar saldo da carteira","description":"Retorna o saldo total disponível de uma carteira com base no ID informado","operationId":"getWalletBalance","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Saldo da carteira retornado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WalletDTO"}}}},"404":{"description":"Carteira não encontrada"},"500":{"description":"Erro interno no servidor"}}}},"/api/v1/users/{id}":{"get":{"tags":["UserController"],"summary":"Buscar usuário por ID","description":"Retorna os dados de um usuário com base no ID fornecido","operationId":"findUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Usuário retornado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}},"404":{"description":"Usuário não encontrado"},"500":{"description":"Erro interno"}}},"delete":{"tags":["UserController"],"summary":"Excluir usuário por ID","description":"Exclui um usuário com base no ID fornecido","operationId":"deleteUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Usuário excluído com sucesso"},"404":{"description":"Usuário não encontrado"},"500":{"description":"Erro interno"}}}},"/api/v1/positions/wallets/{walletId}":{"get":{"tags":["PositionController"],"summary":"Buscar posição de um ativo em uma carteira","description":"Retorna os dados da posição de um ativo específico dentro de uma carteira, com base no ID da carteira e símbolo do ativo","operationId":"getPosition","parameters":[{"name":"walletId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"assetSymbol","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Posição retornada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PositionDTO"}}}},"404":{"description":"Posição não encontrada"},"500":{"description":"Erro interno"}}}},"/api/v1/positions/allpositions/wallets/{walletId}":{"get":{"tags":["PositionController"],"summary":"Retorna todas as posições de uma carteira","description":"Retorna todas as posições de ativos associadas a uma carteira","operationId":"getAllPositions","parameters":[{"name":"walletId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Posições retornadas com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PositionDTO"}}}}},"500":{"description":"Erro interno"}}}},"/api/v1/orders":{"get":{"tags":["OrderController"],"summary":"Listar todas as ordens","description":"Retorna uma lista com todas as ordens","operationId":"getAllOrders","responses":{"200":{"description":"Lista de ordens retornada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDTO"}}}}},"500":{"description":"Erro interno no servidor"}}}},"/api/v1/orders/{id}":{"get":{"tags":["OrderController"],"summary":"Buscar ordem por ID","description":"Retorna os dados completos de uma ordem com base no seu ID","operationId":"findOrderById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Ordem encontrada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}},"404":{"description":"Ordem não encontrada"},"500":{"description":"Erro interno no servidor"}}}},"/api/v1/assets/{symbol}":{"get":{"tags":["AssetController"],"summary":"Buscar cotação de ativo","description":"Retorna os dados de cotação de um ativo com base no seu símbolo","operationId":"getAsset","parameters":[{"name":"symbol","in":"path","description":"Símbolo do ativo(ex: PETR4, ITUB4, VALE3)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Cotação retornada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuoteResponseDTO"}}}},"404":{"description":"Ativo não encontrado"},"500":{"description":"Erro interno"}}}}},"components":{"schemas":{"SellOrderDTO":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"quantityToSell":{"type":"integer","format":"int32"}}},"OrderDTO":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"assetSymbol":{"type":"string"},"walletId":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["BUY","SELL"]},"quantity":{"type":"integer","format":"int32"},"priceAtExecution":{"type":"number"},"tradeTime":{"type":"string","format":"date-time"}}},"UserRequestDTO":{"type":"object","properties":{"name":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"userWalletId":{"type":"integer","format":"int64"}}},"BuyRequestDTO":{"type":"object","properties":{"assetSymbol":{"type":"string","description":"Símbolo do ativo a ser comprado","example":"PETR4"},"quantity":{"type":"integer","format":"int32"}}},"WalletDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"balance":{"type":"number"},"userId":{"type":"integer","format":"int64"}}},"PositionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"walletId":{"type":"integer","format":"int64"},"assetSymbol":{"type":"string"},"totalQuantity":{"type":"integer","format":"int32"},"averagePrice":{"type":"number"}}},"AssetDTO":{"type":"object","properties":{"symbol":{"type":"string"},"shortName":{"type":"string"},"regularMarketPrice":{"type":"number","format":"double"}}},"QuoteResponseDTO":{"type":"object","properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/AssetDTO"}}}}}}}